Added a bootleg reflection system to scene graph
All components that we need to be able to display in the graph will have to implement a void Display(ComponentType component) function; This should be further templatized to support various return values and more input parameters if required. All components that the editor can display will need to be listed in a parameter pack, I want to make this declarative for easier use.